Created a list of products with pagination and filtering by name.
Created a list of products with pagination and filtering by name.
The main form have a DataGrid with the size of the columns adapted and editing disabled only to show the products.
Also a numericUpDown to select the current page and a comboBox to select the max of results to show per page.
On the top of the DataGrid there's a textBox to search by Product Name. (It filters after press the key Enter)

Created ProductFilters used to filter Products on some methods of DataAcces.cs.
Created Product.cs to represent the Products of the database.

Created Helper.cs to get the connectionString from the specified database.

Created DataAcces.cs. Contains a few methods:
GetAllProducts() - returns a List of all the Products.
GetProducts(int limit, int page, ProductFilters filters) - returns a List of Products filtered and paginated.
CountProducts(ProductFilters filters)- returns the count of Products.